requirements.txt
discord.py==2.6.4
python-dotenv==1.2.1
python-escpos==3.1
tzdata==2025.3
